requirements.txt
PySide6>=6.5.0
pyqtgraph>=0.13.0
python-dotenv>=1.0.0
requests>=2.30.0